+++ pkg/test/framework/environments/kubernetes/environment.go
@@ -12,7 +12,7 @@
 //  See the License for the specific language governing permissions and
 //  limitations under the License.
 
-package cluster
+package kubernetes
 
 import (
 	"fmt"
@@ -26,10 +26,10 @@ import (
 
 	"istio.io/istio/pkg/log"
 	"istio.io/istio/pkg/test/dependency"
-	"istio.io/istio/pkg/test/environment"
-	"istio.io/istio/pkg/test/internal"
+	"istio.io/istio/pkg/test/framework"
+	"istio.io/istio/pkg/test/framework/internal"
+	"istio.io/istio/pkg/test/framework/tmpl"
 	"istio.io/istio/pkg/test/kube"
-	"istio.io/istio/pkg/test/tmpl"
 )
 
 var scope = log.RegisterScope("testframework", "General scope for the test framework", 0)
@@ -58,7 +58,7 @@ type Environment struct {
 	TestNamespace string
 }
 
-var _ environment.Interface = &Environment{}
+var _ framework.Environment = &Environment{}
 var _ internal.Environment = &Environment{}
 var _ io.Closer = &Environment{}
 
@@ -154,16 +154,16 @@ func (e *Environment) Reset() error {
 }
 
 // GetMixer returns a deployed Mixer instance in the environment.
-func (e *Environment) GetMixer() (environment.DeployedMixer, error) {
+func (e *Environment) GetMixer() (framework.DeployedMixer, error) {
 	s, err := e.get(dependency.Mixer)
 	if err != nil {
 		return nil, err
 	}
-	return s.(environment.DeployedMixer), nil
+	return s.(framework.DeployedMixer), nil
 }
 
-// GetMixerOrFail returns a deployed Mixer instance in the environment, or fails the test if unsuccessful.
-func (e *Environment) GetMixerOrFail(t testing.TB) environment.DeployedMixer {
+// GetMixerOrFail returns a deployed Mixer instance in the framework, or fails the test if unsuccessful.
+func (e *Environment) GetMixerOrFail(t testing.TB) framework.DeployedMixer {
 	t.Helper()
 
 	m, err := e.GetMixer()
@@ -175,16 +175,16 @@ func (e *Environment) GetMixerOrFail(t testing.TB) environment.DeployedMixer {
 }
 
 // GetPilot returns a deployed Pilot instance in the environment.
-func (e *Environment) GetPilot() (environment.DeployedPilot, error) {
+func (e *Environment) GetPilot() (framework.DeployedPilot, error) {
 	s, err := e.get(dependency.Pilot)
 	if err != nil {
 		return nil, err
 	}
-	return s.(environment.DeployedPilot), nil
+	return s.(framework.DeployedPilot), nil
 }
 
 // GetPilotOrFail returns a deployed Pilot instance in the environment, or fails the test if unsuccessful.
-func (e *Environment) GetPilotOrFail(t testing.TB) environment.DeployedPilot {
+func (e *Environment) GetPilotOrFail(t testing.TB) framework.DeployedPilot {
 	t.Helper()
 
 	m, err := e.GetPilot()
@@ -196,18 +196,18 @@ func (e *Environment) GetPilotOrFail(t testing.TB) environment.DeployedPilot {
 }
 
 // GetAPIServer returns a handle to the ambient API Server in the environment.
-func (e *Environment) GetAPIServer() (environment.DeployedAPIServer, error) {
+func (e *Environment) GetAPIServer() (framework.DeployedAPIServer, error) {
 	a, err := e.get(dependency.APIServer)
 	if err != nil {
 		return nil, err
 	}
 
-	return a.(environment.DeployedAPIServer), nil
+	return a.(framework.DeployedAPIServer), nil
 }
 
 // GetAPIServerOrFail returns a handle to the ambient API Server in the environment, or fails the test if
 // unsuccessful.
-func (e *Environment) GetAPIServerOrFail(t testing.TB) environment.DeployedAPIServer {
+func (e *Environment) GetAPIServerOrFail(t testing.TB) framework.DeployedAPIServer {
 	t.Helper()
 
 	m, err := e.GetAPIServer()
@@ -219,12 +219,12 @@ func (e *Environment) GetAPIServerOrFail(t testing.TB) environment.DeployedAPISe
 }
 
 // GetApp returns a fake testing app object for the given name.
-func (e *Environment) GetApp(name string) (environment.DeployedApp, error) {
+func (e *Environment) GetApp(name string) (framework.DeployedApp, error) {
 	return getApp(name, e.TestNamespace)
 }
 
 // GetAppOrFail returns a fake testing app object for the given name, or fails the test if unsuccessful.
-func (e *Environment) GetAppOrFail(name string, t testing.TB) environment.DeployedApp {
+func (e *Environment) GetAppOrFail(name string, t testing.TB) framework.DeployedApp {
 	t.Helper()
 	a, err := getApp(name, e.TestNamespace)
 	if err != nil {
@@ -234,14 +234,14 @@ func (e *Environment) GetAppOrFail(name string, t testing.TB) environment.Deploy
 }
 
 // GetFortioApp returns a Fortio App object for the given name.
-func (e *Environment) GetFortioApp(name string) (environment.DeployedFortioApp, error) {
+func (e *Environment) GetFortioApp(name string) (framework.DeployedFortioApp, error) {
 	// TODO: Implement or remove the method
 	// See https://github.com/istio/istio/issues/6171
 	panic("Not yet implemented")
 }
 
 // GetFortioAppOrFail returns a Fortio App object for the given name, or fails the test if unsuccessful.
-func (e *Environment) GetFortioAppOrFail(name string, t testing.TB) environment.DeployedFortioApp {
+func (e *Environment) GetFortioAppOrFail(name string, t testing.TB) framework.DeployedFortioApp {
 	t.Helper()
 	a, err := e.GetFortioApp(name)
 	if err != nil {
@@ -251,16 +251,16 @@ func (e *Environment) GetFortioAppOrFail(name string, t testing.TB) environment.
 }
 
 // GetFortioApps returns a set of Fortio Apps based on the given selector.
-func (e *Environment) GetFortioApps(selector string, t testing.TB) []environment.DeployedFortioApp {
+func (e *Environment) GetFortioApps(selector string, t testing.TB) []framework.DeployedFortioApp {
 	// TODO: Implement or remove the method
 	// See https://github.com/istio/istio/issues/6171
 	panic("Not yet implemented")
 }
 
 // GetPolicyBackendOrFail returns the mock policy backend that is used by Mixer for policy checks and reports.
-func (e *Environment) GetPolicyBackendOrFail(t testing.TB) environment.DeployedPolicyBackend {
+func (e *Environment) GetPolicyBackendOrFail(t testing.TB) framework.DeployedPolicyBackend {
 	t.Helper()
-	return e.getOrFail(t, dependency.PolicyBackend).(environment.DeployedPolicyBackend)
+	return e.getOrFail(t, dependency.PolicyBackend).(framework.DeployedPolicyBackend)
 }
 
 func (e *Environment) get(dep dependency.Instance) (interface{}, error) {
